@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
-		<Version>2.2.0-beta1</Version>
+		<Version>3.0.0</Version>
 		<RootNamespace>ToolBX.$(AssemblyName)</RootNamespace>
 		<PackageId>ToolBX.$(AssemblyName)</PackageId>
 		<Product>ToolBX.$(AssemblyName)</Product>
@@ -26,10 +26,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="ToolBX.Mathemancy" Version="2.2.1" />
-		<PackageReference Include="ToolBX.OPEX" Version="2.2.0" />
-		<PackageReference Include="ToolBX.OutWarden" Version="2.2.0" />
-		<PackageReference Include="ToolBX.Reflection4Humans.Extensions" Version="2.2.0" />
+		<PackageReference Include="ToolBX.Mathemancy" Version="3.0.0" />
+		<PackageReference Include="ToolBX.OPEX" Version="3.0.1" />
+		<PackageReference Include="ToolBX.OutWarden" Version="3.0.0" />
+		<PackageReference Include="ToolBX.Reflection4Humans.Extensions" Version="3.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
