@@ -1,20 +1,19 @@
 // Global using directives
 
+global using FluentAssertions;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Newtonsoft.Json;
 global using System;
 global using System.Collections.Generic;
 global using System.IO;
 global using System.Linq;
 global using System.Text.Json;
-global using AutoFixture;
-global using FluentAssertions;
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using Newtonsoft.Json;
 global using ToolBX.Collections.Grid;
 global using ToolBX.Collections.Grid.Json;
 global using ToolBX.Collections.Grid.Resources;
 global using ToolBX.Collections.UnitTesting.Customizations;
+global using ToolBX.Collections.UnitTesting.Extensions;
 global using ToolBX.Eloquentest;
-global using ToolBX.Eloquentest.Customizations;
-global using ToolBX.Eloquentest.Extensions;
 global using ToolBX.Mathemancy;
 global using ToolBX.OPEX;
+global using WhiteJackalStudio.TestTools;
