@@ -1,19 +1,20 @@
 ﻿namespace Collections.Inventory.Tests;
 
 [TestClass]
-public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
+public class StockSearchResultTests : RecordTester<StockSearchResult<GarbageItem>>
 {
     protected override void InitializeTest()
     {
         base.InitializeTest();
         JsonSerializerOptions.WithInventoryConverters();
+        Dummy.Options.MaximumDepth = 4;
     }
 
     [TestMethod]
     public void Indexer_WhenIndexIsWithinBoundaries_ReturnItemAtIndex()
     {
         //Arrange
-        var instance = Fixture.Create<StockSearchResult<IndexedEntry<DummyItem>>>();
+        var instance = Dummy.Create<StockSearchResult<IndexedEntry<GarbageItem>>>();
 
         var index = instance.GetRandomIndex();
 
@@ -30,7 +31,7 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
         //Arrange
 
         //Act
-        var action = () => new StockSearchResult<IndexedEntry<DummyItem>>(null!);
+        var action = () => new StockSearchResult<IndexedEntry<GarbageItem>>(null!);
 
         //Assert
         action.Should().Throw<ArgumentNullException>().WithParameterName("items");
@@ -40,10 +41,10 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void Constructor_WhenItemsIsEmpty_ShouldBeEmpty()
     {
         //Arrange
-        var items = new List<IndexedEntry<DummyItem>>();
+        var items = new List<IndexedEntry<GarbageItem>>();
 
         //Act
-        var result = new StockSearchResult<DummyItem>(items);
+        var result = new StockSearchResult<GarbageItem>(items);
 
         //Assert
         result.Should().BeEmpty();
@@ -53,10 +54,10 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void Constructor_WhenItemsIsNotEmpty_ReturnWithItems()
     {
         //Arrange
-        var items = Fixture.CreateMany<IndexedEntry<DummyItem>>().ToList();
+        var items = Dummy.CreateMany<IndexedEntry<GarbageItem>>().ToList();
 
         //Act
-        var result = new StockSearchResult<DummyItem>(items);
+        var result = new StockSearchResult<GarbageItem>(items);
 
         //Assert
         result.Should().BeEquivalentTo(items);
@@ -66,9 +67,9 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void Enumerator_Always_CorrectlyEnumeratesEveryItem()
     {
         //Arrange
-        var searchResult = Fixture.Create<StockSearchResult<DummyItem>>();
+        var searchResult = Dummy.Create<StockSearchResult<GarbageItem>>();
 
-        var enumeratedItems = new List<IndexedEntry<DummyItem>>();
+        var enumeratedItems = new List<IndexedEntry<GarbageItem>>();
 
         //Act
         foreach (var item in searchResult)
@@ -84,7 +85,7 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void GetEnumerator_Always_ReturnsEnumerator()
     {
         //Arrange
-        var searchResult = Fixture.Create<StockSearchResult<DummyItem>>();
+        var searchResult = Dummy.Create<StockSearchResult<GarbageItem>>();
 
         //Act
         var enumerator = searchResult.GetEnumerator();
@@ -97,7 +98,7 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void InterfaceGetEnumerator_Always_ReturnsInternalCollectionEnumerator()
     {
         //Arrange
-        var searchResult = Fixture.Create<StockSearchResult<DummyItem>>();
+        var searchResult = Dummy.Create<StockSearchResult<GarbageItem>>();
 
         //Act
         var enumerator = ((IEnumerable)searchResult).GetEnumerator();
@@ -110,7 +111,7 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void Group_WhenIsEmpty_ReturnEmpty()
     {
         //Arrange
-        var searchResult = new StockSearchResult<DummyItem>(Array.Empty<IndexedEntry<DummyItem>>());
+        var searchResult = new StockSearchResult<GarbageItem>(Array.Empty<IndexedEntry<GarbageItem>>());
 
         //Act
         var result = searchResult.Group();
@@ -123,34 +124,34 @@ public class StockSearchResultTests : RecordTester<StockSearchResult<DummyItem>>
     public void Group_WhenHasUniqueItems_DoNotGroup()
     {
         //Arrange
-        var searchResult = Fixture.Create<StockSearchResult<DummyItem>>();
+        var searchResult = Dummy.Create<StockSearchResult<GarbageItem>>();
 
         //Act
         var result = searchResult.Group();
 
         //Assert
-        result.Should().BeEquivalentTo(searchResult.Select(x => new GroupedEntry<DummyItem>(x.Item, x.Quantity, new List<int> { x.Index })));
+        result.Should().BeEquivalentTo(searchResult.Select(x => new GroupedEntry<GarbageItem>(x.Item, x.Quantity, [x.Index])));
     }
 
     [TestMethod]
     public void Group_WhenHasDuplicateItems_GroupThem()
     {
         //Arrange
-        var entries = Fixture.CreateMany<IndexedEntry<DummyItem>>().ToList();
-        var duplicate = entries.GetRandom() with { Quantity = Fixture.Create<int>(), Index = Fixture.Create<int>() };
+        var entries = Dummy.CreateMany<IndexedEntry<GarbageItem>>().ToList();
+        var duplicate = entries.GetRandom() with { Quantity = Dummy.Create<int>(), Index = Dummy.Create<int>() };
         entries.Add(duplicate);
 
-        var searchResult = new StockSearchResult<DummyItem>(entries);
+        var searchResult = new StockSearchResult<GarbageItem>(entries);
 
         //Act
         var result = searchResult.Group();
 
         //Assert
-        var expectedGroupedEntry = new GroupedEntry<DummyItem>(duplicate.Item, entries.Where(x => x.Item == duplicate.Item).Sum(x => x.Quantity), entries.Where(x => x.Item.Equals(duplicate.Item)).Select(x => x.Index).ToList());
-        var expectedEntries = searchResult.Where(x => !x.Item.Equals(duplicate.Item)).Select(x => new GroupedEntry<DummyItem>(x.Item, x.Quantity, new List<int> { x.Index })).Append(expectedGroupedEntry).ToList();
+        var expectedGroupedEntry = new GroupedEntry<GarbageItem>(duplicate.Item, entries.Where(x => x.Item == duplicate.Item).Sum(x => x.Quantity), entries.Where(x => x.Item.Equals(duplicate.Item)).Select(x => x.Index).ToList());
+        var expectedEntries = searchResult.Where(x => !x.Item.Equals(duplicate.Item)).Select(x => new GroupedEntry<GarbageItem>(x.Item, x.Quantity, [x.Index])).Append(expectedGroupedEntry).ToList();
         result.Should().BeEquivalentTo(expectedEntries);
     }
 
     [TestMethod]
-    public void Ensure_IsJsonSerializable() => Ensure.IsJsonSerializable<StockSearchResult<IndexedEntry<DummyItem>>>(Fixture, JsonSerializerOptions.WithInventoryConverters());
+    public void Ensure_IsJsonSerializable() => Ensure.IsJsonSerializable<StockSearchResult<IndexedEntry<GarbageItem>>>(Dummy, JsonSerializerOptions.WithInventoryConverters());
 }
