@@ -1,5 +1,4 @@
-﻿global using AutoFixture;
-global using FluentAssertions;
+﻿global using FluentAssertions;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using System;
 global using System.Collections.Generic;
@@ -8,9 +7,4 @@ global using ToolBX.Collections.Common;
 global using ToolBX.Collections.Inventory;
 global using ToolBX.Collections.Inventory.Resources;
 global using ToolBX.Eloquentest;
-global using ToolBX.Eloquentest.Extensions;
-global using System.IO;
-global using System.Xml;
-global using System.Xml.Serialization;
-global using Newtonsoft.Json;
-global using ToolBX.Reflection4Humans.Extensions;
+global using ToolBX.Collections.UnitTesting;
