@@ -2,8 +2,14 @@
 
 public abstract record EntryBase<T>
 {
-    public T Item { get; init; } = default!;
-    public int Quantity { get; init; }
+    public T? Item { get; init; }
+
+    public int Quantity
+    {
+        get => _quantity;
+        init => _quantity = value < 0 ? throw new ArgumentOutOfRangeException(nameof(value), value, Exceptions.QuantityMustBePositive) : value;
+    }
+    private readonly int _quantity = 1;
 
     protected EntryBase()
     {
@@ -16,7 +22,7 @@ public abstract record EntryBase<T>
         Quantity = quantity;
     }
 
-    public void Deconstruct(out T item, out int quantity)
+    public void Deconstruct(out T? item, out int quantity)
     {
         item = Item;
         quantity = Quantity;
