@@ -1,6 +1,6 @@
-﻿namespace Collections.ReadOnly.Tests;
+﻿namespace Collections.ObservableDictionary.Tests;
 
-public record Dummy
+public record Garbage
 {
     public int Id { get; init; }
     public string Description { get; init; } = string.Empty;
