@@ -1,17 +1,16 @@
 // Global using directives
 
+global using FluentAssertions;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;
-global using AutoFixture;
-global using FluentAssertions;
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using ToolBX.Collections.Common;
 global using ToolBX.Collections.ObservableDictionary;
 global using ToolBX.Collections.ObservableDictionary.Json;
 global using ToolBX.Collections.ObservableDictionary.Resources;
 global using ToolBX.Collections.UnitTesting.Extensions;
 global using ToolBX.Eloquentest;
-global using ToolBX.Eloquentest.Extensions;
 global using ToolBX.OPEX;
 global using ToolBX.OutWarden;
+global using WhiteJackalStudio.TestTools;
