@@ -7,7 +7,7 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryKeySelector_WhenSourceIsNull_Throw()
     {
         //Arrange
-        IEnumerable<Dummy> source = null!;
+        IEnumerable<Garbage> source = null!;
 
         //Act
         var action = () => source.ToObservableDictionary(x => x.Id);
@@ -20,8 +20,8 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryKeySelector_WhenKeySelectorIsNull_Throw()
     {
         //Arrange
-        var source = Fixture.Create<IEnumerable<KeyValuePair<int, Dummy>>>();
-        Func<KeyValuePair<int, Dummy>, int> keySelector = null!;
+        var source = Dummy.Create<IEnumerable<KeyValuePair<int, Garbage>>>();
+        Func<KeyValuePair<int, Garbage>, int> keySelector = null!;
 
         //Act
         var action = () => source.ToObservableDictionary(keySelector);
@@ -34,7 +34,7 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryKeySelector_WhenSourceIsEmpty_ReturnEmpty()
     {
         //Arrange
-        var source = Enumerable.Empty<Dummy>();
+        var source = Enumerable.Empty<Garbage>();
 
         //Act
         var result = source.ToObservableDictionary(x => x.Id);
@@ -47,7 +47,7 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryKeySelector_WhenSourceIsNotEmpty_ReturnEquivalentDictionary()
     {
         //Arrange
-        var source = Fixture.CreateMany<Dummy>().ToList();
+        var source = Dummy.CreateMany<Garbage>().ToList();
 
         //Act
         var result = source.ToObservableDictionary(x => x.Id);
@@ -60,9 +60,9 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryBothSelectors_WhenSourceIsNull_Throw()
     {
         //Arrange
-        IEnumerable<KeyValuePair<int, Dummy>> source = null!;
-        var keySelector = Fixture.Create<Func<KeyValuePair<int, Dummy>, int>>();
-        var elementSelector = Fixture.Create<Func<KeyValuePair<int, Dummy>, Dummy>>();
+        IEnumerable<KeyValuePair<int, Garbage>> source = null!;
+        var keySelector = Dummy.Create<Func<KeyValuePair<int, Garbage>, int>>();
+        var elementSelector = Dummy.Create<Func<KeyValuePair<int, Garbage>, Garbage>>();
 
         //Act
         var action = () => source.ToObservableDictionary(keySelector, elementSelector);
@@ -75,9 +75,9 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryBothSelectors_WhenKeySelectorIsNull_Throw()
     {
         //Arrange
-        var source = Fixture.Create<IEnumerable<KeyValuePair<int, Dummy>>>();
-        Func<KeyValuePair<int, Dummy>, int> keySelector = null!;
-        var elementSelector = Fixture.Create<Func<KeyValuePair<int, Dummy>, Dummy>>();
+        var source = Dummy.Create<IEnumerable<KeyValuePair<int, Garbage>>>();
+        Func<KeyValuePair<int, Garbage>, int> keySelector = null!;
+        var elementSelector = Dummy.Create<Func<KeyValuePair<int, Garbage>, Garbage>>();
 
         //Act
         var action = () => source.ToObservableDictionary(keySelector, elementSelector);
@@ -90,9 +90,9 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryBothSelectors_WhenElementSelectorIsNull_Throw()
     {
         //Arrange
-        var source = Fixture.Create<IEnumerable<KeyValuePair<int, Dummy>>>();
-        var keySelector = Fixture.Create<Func<KeyValuePair<int, Dummy>, int>>();
-        Func<KeyValuePair<int, Dummy>, Dummy> elementSelector = null!;
+        var source = Dummy.Create<IEnumerable<KeyValuePair<int, Garbage>>>();
+        var keySelector = Dummy.Create<Func<KeyValuePair<int, Garbage>, int>>();
+        Func<KeyValuePair<int, Garbage>, Garbage> elementSelector = null!;
 
         //Act
         var action = () => source.ToObservableDictionary(keySelector, elementSelector);
@@ -105,7 +105,7 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryBothSelectors_WhenSourceIsEmpty_ReturnEmpty()
     {
         //Arrange
-        var source = Enumerable.Empty<KeyValuePair<int, Dummy>>();
+        var source = Enumerable.Empty<KeyValuePair<int, Garbage>>();
 
         //Act
         var result = source.ToObservableDictionary(x => x.Key, x => x.Value);
@@ -118,7 +118,7 @@ public sealed class ObservableDictionaryExtensionsTests : Tester
     public void ToObservableDictionaryBothSelectors_WhenSourceIsNotEmpty_ReturnEquivalentDictionary()
     {
         //Arrange
-        var source = Fixture.Create<Dictionary<int, Dummy>>();
+        var source = Dummy.Create<Dictionary<int, Garbage>>();
 
         //Act
         var result = source.ToObservableDictionary(x => x.Key, x => x.Value);
