@@ -1,20 +1,11 @@
 // Global using directives
 
-global using System;
-global using System.Collections;
-global using System.Collections.Generic;
-global using System.IO;
-global using System.Linq;
-global using System.Xml;
-global using System.Xml.Serialization;
-global using AutoFixture;
 global using FluentAssertions;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using Newtonsoft.Json;
-global using ToolBX.Collections.Common;
+global using System;
+global using System.Collections.Generic;
+global using System.Linq;
 global using ToolBX.Collections.ObservableList;
-global using ToolBX.Collections.ObservableList.Resources;
 global using ToolBX.Collections.UnitTesting.ObservableList;
 global using ToolBX.Eloquentest;
-global using ToolBX.Eloquentest.Extensions;
-global using ToolBX.OPEX;
+global using WhiteJackalStudio.TestTools;
