@@ -1,10 +1,10 @@
 ﻿namespace Collections.ReadOnly.Tests;
 
-public sealed record DummyWithListInside
+public sealed record GarbageWithListInside
 {
     public Guid Id { get; init; }
 
     public string Name { get; init; } = string.Empty;
 
-    public ReadOnlyList<Dummy> Stuff { get; init; } = [];
+    public ReadOnlyList<Garbage> Stuff { get; init; } = [];
 }
