@@ -1,13 +1,17 @@
 // Global using directives
 
+global using FluentAssertions;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using System;
 global using System.Collections.Generic;
 global using System.Collections.Immutable;
 global using System.Linq;
 global using System.Text.Json;
-global using AutoFixture;
-global using FluentAssertions;
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Newtonsoft.Json;
 global using ToolBX.Collections.ReadOnly;
 global using ToolBX.Collections.ReadOnly.Json;
+global using ToolBX.Collections.UnitTesting.Extensions;
+global using ToolBX.Dummies;
 global using ToolBX.Eloquentest;
+global using ToolBX.Eloquentest.Dummies;
+global using WhiteJackalStudio.TestTools;
