@@ -1,19 +1,8 @@
 ﻿namespace ToolBX.Collections.UnitTesting.Customizations;
 
-public class ReadOnlyListSpecimenBuilder : ISpecimenBuilder
+public sealed class ReadOnlyListCustomization : ListCustomizationBase
 {
-    public object Create(object request, ISpecimenContext context)
-    {
-        if (request is Type type && type.IsGenericType && type.GetGenericTypeDefinition() == typeof(ReadOnlyList<>))
-        {
-            var elementType = type.GetGenericArguments()[0];
-            var listType = typeof(List<>).MakeGenericType(elementType);
-            var list = context.Resolve(listType);
+    protected override IEnumerable<Type> Types { get; } = [typeof(ReadOnlyList<>)];
 
-            var create = typeof(ReadOnlyListExtensions).GetSingleMethod(nameof(ReadOnlyListExtensions.ToReadOnlyList)).MakeGenericMethod(elementType);
-            return create.Invoke(null, new[] { list })!;
-        }
-
-        return new NoSpecimen();
-    }
+    protected override object Convert<T>(IEnumerable<T> source) => source.ToReadOnlyList();
 }
