@@ -1,11 +1,11 @@
 // Global using directives
 
-global using System.Collections;
-global using System.Text.Json;
-global using AutoFixture;
-global using AutoFixture.Kernel;
 global using FluentAssertions;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Moq;
+global using System.Collections;
+global using System.Reflection;
+global using System.Text.Json;
 global using ToolBX.Collections.Caching;
 global using ToolBX.Collections.Common;
 global using ToolBX.Collections.Grid;
@@ -18,8 +18,12 @@ global using ToolBX.Collections.ObservableList.Resources;
 global using ToolBX.Collections.ObservableStack;
 global using ToolBX.Collections.ReadOnly;
 global using ToolBX.Collections.UnitTesting.Extensions;
+global using ToolBX.Dummies;
+global using ToolBX.Dummies.Customizations;
 global using ToolBX.Eloquentest;
-global using ToolBX.Eloquentest.Customizations;
-global using ToolBX.Eloquentest.Extensions;
+global using ToolBX.Eloquentest.Dummies;
+global using ToolBX.Mathemancy;
 global using ToolBX.OPEX;
 global using ToolBX.Reflection4Humans.Extensions;
+global using ToolBX.Reflection4Humans.TypeFetcher;
+global using WhiteJackalStudio.TestTools;
