@@ -1,6 +1,6 @@
 ﻿namespace ToolBX.Collections.UnitTesting.Inventory;
 
-public record DummyItem
+public record GarbageItem
 {
     public int Id { get; init; }
     public string Name { get; init; } = string.Empty;
